MAJOR CHANGES: Core of Eddie is now multi-threaded using a scheduler thread to run each check in its own thread.  Thread usage is limited so things don't get out of control.
The scheduler tracks jobs with a derivative of Python's Queue class which orders
items by time, so that the job to be started soonest will be at the front
of the queue.  This will now allow directives to specify their own scanperiod
and execute as often or as little as desired, indepentently of other directives.

The config parser does not support scanperiod specification for each directive yet,
but will soon.

The limit of threads to spawn is hard-coded in eddie.py but will soon be a config
variable in eddie.cf.

Modified config files are still automatically detected (sometime within a 10 minute
period by the "Housecleaning" thread (main process)) which causes the scheduler
to be signalled to exit and then the configs are re-read and a new scheduler will
be started up.

Signal handling is still a bit iffy, especially SIGHUP which is supposed to
force a config re-read.  This does not work properly yet and shouldn't be used.

NOTE: this has only been tested under Linux so far, but will be tested under
Solaris soon.  Threading relies on the Python Threading module which is
supposed to be portable.