timeQueue is the queueing class derived from Python's Queue class.  It is as thread-friendly as Queue, the major difference being objects are inserted into the queue based on a given time.  Objects with the lowest times are closest to the front of the queue.
To support this, objects have to be added along with their time, so a
2-tuple must be added, eg: q.put( (obj, time) ).  Similarly q.get()
returns the same 2-tuple.

An extra public method has been added, over what Queue offers, q.head().
This method returns the item (and time) from the front of the queue,
exactly as q.get(), but does not remove it from the queue.