New config format, which is not compatable with old format. All arguments to a directive are now named arguments.
Eg: a PROC directive that looked like this:
    PROC inetd:     NR      COMMONALERT(commonmsg.proc,2)
will now look like:
    PROC 'inetd':   procname='inetd' rule='NR'      action="COMMONALERT(commonmsg.proc,1)"

The string after the directive name and before the ':' should be
a unique name for that particular directive.  This will allow
other directives to refer back to it for dependencies, etc, when
they are implemented.

This change was made to support adding new arguments to directives
for more powerful configurations.

Note that if the value of the argument is more than simply alphanumeric
characters then quotes will be required.  In the example above, quotes
can actually be safely left out for 'inetd' and 'NR', but not for the
action value.

This change has currently been made to all common directives in directives.py
but has not yet been made to directives in the Directives/ directory.