Added sockets.py - A sockets interface to the current state of all eddie checks, this will be used for a console like interface.
Added config var CONSPORT - this is the port to listen to
console connections on.  The default is 33343 or eddie on the
telephone keypad.

Chaged eddie.py - changes include cleaning up the way threads
are started and stoped, there is now start_threads() and
stop_threads().  I did this so that both the scheduler thread
and the console socket thread can be started and stop easily
when the config changes.

NOTED BUGS - Needs a catch in SIGHUP to catch the interuppted
system call - severity minor