Enforced unique group and directive names at same group level.
Implemented check and action dependency definitions.  Two new directive
options are: actiondependson and checkdependson.  These can be set to a
string containing a list of directives (comma-separated) that this directive
is dependent on.  If any of the dependent directives has failed when this
directive comes to perform its check or action (depending on which option
was used) then that check or action will be skipped.

Added new directive option actionperiod.  This is a string containing an
expression which, when evaluated, sets the current period between actions
being performed.  This allows for periods between actions to different to
the period between checks.  It also allows for the period to be defined by
a mathematical expression, so the action period could exponentially increase
for example (for actions called during a single failure - the action period
will be reset when the failure is fixed).

Improved error handling of console connections from bad clients.